refactor: extract duplicated SyncHistoryDetailSheet into shared component
