+++ frontend/plugins/mongolian_ui/src/modules/msdynamic/msdynamic-sync-history/components/stringifySyncValue.ts
@@ -1,27 +1,4 @@
-export const stringifySyncValue = (value: unknown) => {
-  if (!value) {
-    return '';
-  }
-
-  if (typeof value === 'string') {
-    try {
-      return JSON.stringify(JSON.parse(value), null, 2);
-    } catch {
-      return value;
-    }
-  }
-
-  return JSON.stringify(value, null, 2);
-};
-
-export const stringifySyncValueInline = (value: unknown) => {
-  if (!value) {
-    return '';
-  }
-
-  if (typeof value === 'string') {
-    return value;
-  }
-
-  return JSON.stringify(value);
-};
+export {
+  stringifySyncValue,
+  stringifySyncValueInline,
+} from '@/shared/sync-history/components/stringifySyncValue';
