+++ frontend/plugins/mongolian_ui/src/modules/shared/sync-history/components/stringifySyncValue.ts
@@ -1,0 +1,27 @@
+export const stringifySyncValue = (value: unknown) => {
+  if (!value) {
+    return '';
+  }
+
+  if (typeof value === 'string') {
+    try {
+      return JSON.stringify(JSON.parse(value), null, 2);
+    } catch {
+      return value;
+    }
+  }
+
+  return JSON.stringify(value, null, 2);
+};
+
+export const stringifySyncValueInline = (value: unknown) => {
+  if (!value) {
+    return '';
+  }
+
+  if (typeof value === 'string') {
+    return value;
+  }
+
+  return JSON.stringify(value);
+};
